+++ src/Commands/Apps/AddApp.cs
@@ -1,11 +1,14 @@
-﻿using PnP.Framework.ALM;
+﻿using Microsoft.Online.SharePoint.TenantAdministration;
+using Microsoft.SharePoint.Client;
+using PnP.Framework.ALM;
 using PnP.Framework.Enums;
+using PnP.PowerShell.Commands.Base;
 using System.Management.Automation;
 
 namespace PnP.PowerShell.Commands.Apps
 {
     [Cmdlet(VerbsCommon.Add, "PnPApp")]
-    public class AddApp : PnPSharePointCmdlet
+    public class AddApp : PnPAdminCmdlet
     {
         [Parameter(Mandatory = true, Position = 0, ValueFromPipeline = true)]
         public string Path;
@@ -25,8 +28,35 @@ namespace PnP.PowerShell.Commands.Apps
         [Parameter(Mandatory = false)]
         public int Timeout = 200;
 
+        [Parameter(Mandatory = false)]
+        public SwitchParameter Force;
+
         protected override void ExecuteCmdlet()
         {
+            bool isScriptSettingUpdated = false;
+
+            if (Scope == AppCatalogScope.Tenant)
+            {
+                var appcatalogUri = ClientContext.Web.GetAppCatalog();
+                var ctx = ClientContext.Clone(appcatalogUri);
+                WriteVerbose("Checking if the tenant app catalog is a no-script site");
+                if (ctx.Site.IsNoScriptSite())
+                {
+                    if (Force || ShouldContinue("The tenant appcatalog is a no-script site. Do you want to temporarily enable scripting on it?", Properties.Resources.Confirm))
+                    {
+                        WriteVerbose("Temporarily enabling scripting on the tenant app catalog site");
+                        var tenant = new Tenant(AdminContext);
+                        tenant.SetSiteProperties(appcatalogUri.AbsoluteUri, noScriptSite: false);
+                        isScriptSettingUpdated = true;
+                    }
+                    else
+                    {
+                        WriteWarning("Scripting is disabled on the tenant app catalog site. This command cannot proceed without allowing scripts.");
+                        return;
+                    }
+                }
+            }
+
             if (!System.IO.Path.IsPathRooted(Path))
             {
                 Path = System.IO.Path.Combine(SessionState.Path.CurrentFileSystemLocation.Path, Path);
@@ -42,7 +72,6 @@ namespace PnP.PowerShell.Commands.Apps
 
             try
             {
-
                 if (Publish)
                 {
                     if (manager.Deploy(result, SkipFeatureDeployment, Scope))
@@ -57,6 +86,18 @@ namespace PnP.PowerShell.Commands.Apps
                 // Exception occurred rolling back
                 manager.Remove(result, Scope);
                 throw;
+            }
+            finally
+            {
+                if (isScriptSettingUpdated)
+                {
+                    WriteVerbose("Disabling scripting on the tenant app catalog site");
+                    var appcatalogUri = ClientContext.Web.GetAppCatalog();
+                    var ctx = ClientContext.Clone(appcatalogUri);
+
+                    var tenant = new Tenant(AdminContext);
+                    tenant.SetSiteProperties(appcatalogUri.AbsoluteUri, noScriptSite: true);
+                }
             }
         }
     }