+++ src/Commands/Base/Completers/ContentTypeCompleter.cs
@@ -4,7 +4,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Management.Automation;
 using System.Management.Automation.Language;
-using System.Reflection.Metadata;
 using Microsoft.SharePoint.Client;
 
 namespace PnP.PowerShell.Commands.Base.Completers
@@ -13,8 +12,8 @@ namespace PnP.PowerShell.Commands.Base.Completers
     {
         protected override IEnumerable<CompletionResult> GetArguments(string commandName, string parameterName, string wordToComplete, CommandAst commandAst, IDictionary fakeBoundParameters)
         {
-            IEnumerable<ContentType> result = PnPConnection.Current.Context.LoadQuery(PnPConnection.Current.Context.Web.AvailableContentTypes.Include(f => f.Name));
-            PnPConnection.Current.Context.ExecuteQueryRetry();
+            IEnumerable<ContentType> result = ClientContext.LoadQuery(ClientContext.Web.AvailableContentTypes.Include(f => f.Name));
+            ClientContext.ExecuteQueryRetry();
             foreach (var ct in result.Where(l => l.Name.StartsWith(wordToComplete, StringComparison.InvariantCultureIgnoreCase)))
             {
                 yield return new CompletionResult(ct.Name);