+++ src/Commands/Model/TokenCacheConfiguration.cs
@@ -1,0 +1,9 @@
+namespace PnP.PowerShell.Commands.Model
+{
+    public class TokenCacheConfiguration
+    {
+        public string Url { get; set; }
+        public string ClientId { get; set; }
+        public bool Enabled { get; set; }
+    }
+}